+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - jira: new `debug` option.
+- github: new action `createRepo`.